Phase 1 Complete: File splitting for timeline components
✅ Phase 1: File Compliance (Required)
- [x] Split YearView.tsx → YearView.tsx (315 lines) + hooks/useYearView.ts (291 lines)
- [x] Split MonthView.tsx → MonthView.tsx (321 lines) + hooks/useMonthView.ts (390 lines)
- [x] All components pure (props-only), all logic in hooks
- [x] No file exceeds 400 lines

Changes:
- Created useYearView.ts hook with all YearView logic
- Created useMonthView.ts hook with all MonthView logic
- Refactored YearView.tsx to pure component (removed 217 lines)
- Refactored MonthView.tsx to pure component (removed 304 lines)
- Fixed TypeScript linter errors
- All files now comply with 400-line limit
